+++ sweat/models/exceptions.py
