+++ sweat/models/mixins.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .exceptions import WorkoutDataFrameValidationException
+from sweat.models.exceptions import WorkoutDataFrameValidationException
 
 VALIDATION_COLUMNS = ['time', 'power', 'heartrate', 'cadence', 'speed']
